add alert when you delete a product

+++ src/components/Product.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 //REDUX
 import { useDispatch } from 'react-redux';
@@ -11,7 +12,19 @@ const Product = ({ product }) => {
 
     //Confirm if wants to delete product
     const deleteProductConfirm = id => {
-        dispatch(deleteProductAction(id));
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+        }).then(result => {
+            if (result.isConfirmed) {
+                dispatch(deleteProductAction(id));
+            }
+        });
     };
     return (
         <tr>
